traffic/lambda/dev_traffic_AL_to_raw/requirements.txt
boto3
pandas
numpy
mysql-connector
requests
s3fs
pyyaml